a big big commit

+++ Rng.cpp
@@ -5,7 +5,7 @@ void Rng::init_genrand64(unsigned long long seed)
 {
     mt[0] = seed;
     for (mti=1; mti<NN; mti++) 
-        mt[mti] =  (6364136223846793005ULL * (mt[mti-1] ^ (mt[mti-1] >> 62)) + mti);
+        mt[mti] =  (6364136223846793005ULL * (mt[mti-1] ^ (mt[mti-1] >> 62)) + (unsigned int) mti);
 }
 
 /* generates a random number on [0, 2^64-1]-interval */
@@ -55,19 +55,19 @@ long long Rng::genrand64_int63(void)
 /* generates a random number on [0,1]-real-interval */
 double Rng::genrand64_real1(void)
 {
-    return (genrand64_int64() >> 11) * (1.0/9007199254740991.0);
+    return (double) (genrand64_int64() >> 11) * (1.0/9007199254740991.0);
 }
 
 /* generates a random number on [0,1)-real-interval */
 double Rng::genrand64_real2(void)
 {
-    return (genrand64_int64() >> 11) * (1.0/9007199254740992.0);
+    return (double) (genrand64_int64() >> 11) * (1.0/9007199254740992.0);
 }
 
 /* generates a random number on (0,1)-real-interval */
 double Rng::genrand64_real3(void)
 {
-    return ((genrand64_int64() >> 12) + 0.5) * (1.0/4503599627370496.0);
+    return ((double) (genrand64_int64() >> 12) + 0.5) * (1.0/4503599627370496.0);
 }
 
 //constructeur
@@ -76,7 +76,7 @@ Rng::Rng(unsigned long long seed)
 	mti=NN+1; 
 	mt[0] = seed;
     for (mti=1; mti<NN; mti++) 
-        mt[mti] =  (6364136223846793005ULL * (mt[mti-1] ^ (mt[mti-1] >> 62)) + mti);
+        mt[mti] =  (6364136223846793005ULL * (mt[mti-1] ^ (mt[mti-1] >> 62)) + (unsigned int) mti);
 }
 
 //constructeur
